B&B class updated. - Removed meaningless const.
Handler container class:
- Now min_value_found_in_obj is double.

+++ ParallelBaB/HandlerContainer.cpp
@@ -186,7 +186,7 @@ void HandlerContainer::updateMinValueFound(const Solution &solution) {
             min_value_found_in_obj[objective] = solution.getObjective(objective);
 }
 
-int HandlerContainer::getBestValueFoundIn(int obj) const {
+double HandlerContainer::getBestValueFoundIn(int obj) const {
     return min_value_found_in_obj[obj];
 }
